// src/app/lib/generated/models/HairColor.ts
/* tslint:disable */
/* eslint-disable */
/**
 * AI Babe Public API
 * Lorem ipsum
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const HairColor = {
    Blonde: 'blonde',
    Brunette: 'brunette',
    Black: 'black',
    Pink: 'pink',
    Redhead: 'redhead',
    Blue: 'blue'
} as const;
export type HairColor = typeof HairColor[keyof typeof HairColor];


export function instanceOfHairColor(value: any): boolean {
    for (const key in HairColor) {
        if (Object.prototype.hasOwnProperty.call(HairColor, key)) {
            if (HairColor[key as keyof typeof HairColor] === value) {
                return true;
            }
        }
    }
    return false;
}

export function HairColorFromJSON(json: any): HairColor {
    return HairColorFromJSONTyped(json, false);
}

export function HairColorFromJSONTyped(json: any, ignoreDiscriminator: boolean): HairColor {
    return json as HairColor;
}

export function HairColorToJSON(value?: HairColor | null): any {
    return value as any;
}

export function HairColorToJSONTyped(value: any, ignoreDiscriminator: boolean): HairColor {
    return value as HairColor;
}