// src/app/lib/generated/models/HairStyle.ts
/* tslint:disable */
/* eslint-disable */
/**
 * AI Babe Public API
 * Lorem ipsum
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const HairStyle = {
    Short: 'short',
    Braids: 'braids',
    Curly: 'curly',
    Straight: 'straight',
    Bun: 'bun',
    Ponytail: 'ponytail'
} as const;
export type HairStyle = typeof HairStyle[keyof typeof HairStyle];


export function instanceOfHairStyle(value: any): boolean {
    for (const key in HairStyle) {
        if (Object.prototype.hasOwnProperty.call(HairStyle, key)) {
            if (HairStyle[key as keyof typeof HairStyle] === value) {
                return true;
            }
        }
    }
    return false;
}

export function HairStyleFromJSON(json: any): HairStyle {
    return HairStyleFromJSONTyped(json, false);
}

export function HairStyleFromJSONTyped(json: any, ignoreDiscriminator: boolean): HairStyle {
    return json as HairStyle;
}

export function HairStyleToJSON(value?: HairStyle | null): any {
    return value as any;
}

export function HairStyleToJSONTyped(value: any, ignoreDiscriminator: boolean): HairStyle {
    return value as HairStyle;
}